first attempt at EV bid curve
Making a demand curve based on how long there is left until the car needs to be charged and how long it will take to be charged.

So if the car will take 2 hours to charge from current point. There is 6 hours until it needs to be charged. So the bid curve will be willing to pay a third of the way up the price range.

This logic will most certainly change.